requirements.txt
openai
torch
streamlit
Flask
transformers
flask-cors
accelerate
llama-cpp-python

PyMuPDF
nltk
pandas
openpyxl
gensim
sentence_transformers

chromadb

langchain
langchain-community
langchain-core
langchain_huggingface
huggingface_hub
